Once receiving ConstraintViolationException by default validation annotation information is lost.
